use size_t instead of unsigned int.

+++ include/dmm.h
@@ -1,8 +1,10 @@
 #ifndef DUX_DMM_H
 #define DUX_DMM_H
 
-void dmm_init(unsigned int _kernel_end, unsigned int _end_memory);
-void *kmalloc(unsigned int nbytes);
+#include <stddef.h>
+
+void dmm_init(size_t _kernel_end, size_t _end_memory);
+void *kmalloc(size_t nbytes);
 void kfree(void *ap);
 
 #endif
