+++ src/paging.c
@@ -1,4 +1,5 @@
 #include <dmm.h>
+#include <stddef.h>
 #include "paging.h"
 
 PageDirEntry *cur_page_directory;
@@ -8,7 +9,7 @@ PageDirEntry *cur_page_directory;
  */
 void clear_page_directory(PageDirEntry* page_directory)
 {
-    unsigned int i;
+    size_t i;
     for (i = 0; i < 1024; i++) {
         page_directory[i].present = 0;
         page_directory[i].readwrite = 1;
@@ -19,9 +20,9 @@ void clear_page_directory(PageDirEntry* page_directory)
 /* idmap_page_table(page_table, address)
  * Identity maps the page table given to address.
  */
-void idmap_page_table(PageTableEntry* page_table, unsigned int address)
+void idmap_page_table(PageTableEntry* page_table, size_t address)
 {
-    unsigned int i;
+    size_t i;
     for (i = 0; i < 1024; i++, address += 0x1000) {
         page_table[i].address = address >> 12;
         page_table[i].present = 1;
@@ -43,14 +44,14 @@ void dmm_paging_init()
     idmap_page_table(page_table, 0);
 
     // Add that page table to the page directory.
-    cur_page_directory[0].address = (unsigned int) page_table >> 12;
+    cur_page_directory[0].address = (size_t) page_table >> 12;
     cur_page_directory[0].present = 1;
     cur_page_directory[0].readwrite = 1;
     cur_page_directory[0].user = 0;
 
     // Enable paging.
     __asm__ volatile("mov %0, %%cr3":: "b"(cur_page_directory));
-    unsigned int cr0;
+    size_t cr0;
     __asm__ volatile("mov %%cr0, %0": "=b"(cr0));
     cr0 |= 0x80000000;
     __asm__ volatile("mov %0, %%cr0":: "b"(cr0));
