+++ src/paging.h
@@ -2,37 +2,38 @@
 #define DMM_PAGING_H
 
 #include "placement_allocator.h"
+#include <stddef.h>
 
 typedef struct {
-	unsigned int present : 1;
-	unsigned int readwrite : 1;
-	unsigned int user : 1;
-	unsigned int writethrough : 1;
-	unsigned int cachedisable : 1;
-	unsigned int accessed : 1;
-	unsigned int dirty : 1;
-	unsigned int attributeindex : 1;
-	unsigned int global : 1;
-	unsigned int reserved : 3;
-	unsigned int address : 20;
+	size_t present : 1;
+	size_t readwrite : 1;
+	size_t user : 1;
+	size_t writethrough : 1;
+	size_t cachedisable : 1;
+	size_t accessed : 1;
+	size_t dirty : 1;
+	size_t attributeindex : 1;
+	size_t global : 1;
+	size_t reserved : 3;
+	size_t address : 20;
 } PageTableEntry;
 
 typedef struct {
-	unsigned int present : 1;
-	unsigned int readwrite : 1;
-	unsigned int user : 1;
-	unsigned int writethrough : 1;
-	unsigned int cachedisable : 1;
-	unsigned int accessed : 1;
-	unsigned int available : 1;
-	unsigned int pagesize : 1;
-	unsigned int global : 1;
-	unsigned int reserved : 3;
-	unsigned int address : 20;
+	size_t present : 1;
+	size_t readwrite : 1;
+	size_t user : 1;
+	size_t writethrough : 1;
+	size_t cachedisable : 1;
+	size_t accessed : 1;
+	size_t available : 1;
+	size_t pagesize : 1;
+	size_t global : 1;
+	size_t reserved : 3;
+	size_t address : 20;
 } PageDirEntry;
 
 void clear_page_directory(PageDirEntry* page_directory);
-void idmap_page_table(PageTableEntry* page_table, unsigned int address);
+void idmap_page_table(PageTableEntry* page_table, size_t address);
 void dmm_paging_init();
 
 #endif
